+++ packages/logger/vitest.config.ts
@@ -1,9 +1,4 @@
 import { defineProject, mergeConfig } from 'vitest/config';
 import configShared from '../../vitest.shared.js';
 
-export default mergeConfig(
-	configShared,
-	defineProject({
-		esbuild: { target: 'es2021' }
-	})
-);
+export default mergeConfig(configShared, defineProject({}));