+++ packages/http-framework-test-utils/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @wolfstar/http-framework-test-utils
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies []:
+    - @wolfstar/http-framework@3.0.0
+
+## 2.0.1
+
+### Patch Changes
+
+- [#59](https://github.com/wolfstar-project/stars-components/pull/59) [`b9be51e`](https://github.com/wolfstar-project/stars-components/commit/b9be51ef038beaa1169cf43e4507b1ad2f3ad9db) - Add provenance attestation to publishConfig for all packages
+- Updated dependencies [[`b9be51e`](https://github.com/wolfstar-project/stars-components/commit/b9be51ef038beaa1169cf43e4507b1ad2f3ad9db)]:
+    - @wolfstar/http-framework@2.3.1
+
 ## 2.0.0
 
 ### Major Changes